requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0